requirements.txt
flask
gradio
torch
transformers
Pillow
requests
timm
